+++ js/widgets/calendar.js
@@ -21,6 +21,7 @@ $.widget("metro.calendar", {
         locale: 'en',
         actions: true,
         condensedGrid: false,
+        scheme: 'default',
         getDates: function (d) { },
         dayClick: function (d, d0) { }
     },
@@ -103,6 +104,10 @@ $.widget("metro.calendar", {
             $.each(dates, function () {
                 if (new Date(this) !== undefined) { that.setDateStored(this); }
             });
+        }
+
+        if (o.scheme !== 'default') {
+            element.addClass(o.scheme);
         }
 
         this._renderCalendar();