+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.0.12
 * add color schemes for calendar
+* fix calendar min, max date
 
 ## 3.0.11
 * add keypad onChange event