+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * add color schemes for calendar
 * fix calendar min, max date
 * init clock plugin
+* fix bug in calendar after merge PR 785
 
 ## 3.0.11
 * add keypad onChange event